Add placeholder page for redundant article #26
Article #26 was removed while old site was active and didn't have a
page on the new website. However as there may still be links to this
page in existance the number can't be re-used by a new article.

Therefore this dummy page has been added to be displayed only when
linked to externally. It has been given index:false property to
exclude it from the articles index page.