Restyle and edit main index page.
The home page was restyled so that each summary item on the page could
be clicked to trigger a link to the associated page. Previously these
items had a "More info" link, which was removed.

The page content was edited to suit the new layout. In particular the
new design does not allow links within each text block, so the only
such link was removed.