Enable downloading of codelib releases from GitHub
Modified codelib-tplt.html to use the new `use-github-for-downloads`
page variable to decide whether to make the download button in the
page header bar link to the current release page on GitHub instead of
a SourceForge direct download.

The SourceForge download remains the default if
`use-github-for-downloads` is set to false or is undefined.
The GitHub link is used when `use-github-for-downloads` is set to true.